Fix implicit size conversion

+++ include/gdtoaimp.h
@@ -487,9 +487,9 @@ Bigint {
 typedef unsigned int size_t;
 #endif
 extern void memcpy_D2A ANSI((void*, const void*, size_t));
-#define Bcopy(x,y) memcpy_D2A(&x->sign,&y->sign,y->wds*sizeof(ULong) + 2*sizeof(int))
+#define Bcopy(x,y) memcpy_D2A(&x->sign,&y->sign,(size_t)(y->wds)*sizeof(ULong) + 2*sizeof(int))
 #else /* !NO_STRING_H */
-#define Bcopy(x,y) memcpy(&x->sign,&y->sign,y->wds*sizeof(ULong) + 2*sizeof(int))
+#define Bcopy(x,y) memcpy(&x->sign,&y->sign,(size_t)(y->wds)*sizeof(ULong) + 2*sizeof(int))
 #endif /* NO_STRING_H */
 
 #define Balloc Balloc_D2A
